Morello: Amend definition of (u)intcap_t types
Remove unnecessary Morello ifdef.
Use correct cast in mmap to avoid compiler warning.

+++ src/mman/mmap.c
@@ -31,7 +31,7 @@ void *__mmap(void *start, size_t len, int prot, int flags, int fd, off_t off)
 	ret = __syscall(SYS_mmap, start, len, prot, flags, fd, off);
 #endif
 	/* Fixup incorrect EPERM from kernel. */
-	if ((int)ret == -EPERM && !start && (flags&MAP_ANON) && !(flags&MAP_FIXED))
+	if ((intptr_t)ret == -EPERM && !start && (flags&MAP_ANON) && !(flags&MAP_FIXED))
 		ret = (void *)-ENOMEM;
 	return (void *)__syscall_ret(ret);
 }
